+++ include/UDRefl/FieldPtr.h
@@ -159,8 +159,8 @@ namespace Ubpa::UDRefl {
 		constexpr bool IsConst()    const noexcept { return data.index() & 1; }
 		constexpr bool IsVariable() const noexcept { return !IsConst(); }
 
-		constexpr bool IsOffsetor() const noexcept { return data.index() < 4; }
-		constexpr bool IsObject()   const noexcept { return data.index() >= 4; }
+		constexpr bool IsOwned() const noexcept { return data.index() < 4; }
+		constexpr bool IsUnowned()   const noexcept { return data.index() >= 4; }
 
 		// { variable | const } object
 		ConstObjectPtr RVar() const;