+++ include/UDRefl/IDRegistry.h
@@ -304,8 +304,10 @@ namespace Ubpa::UDRefl {
 
 		TypeIDRegistry();
 
-		using IDRegistry<TypeID>::Register;
-		using IDRegistry<TypeID>::IsRegistered;
+		void   RegisterUnmanaged(TypeID ID, std::string_view name);
+		TypeID RegisterUnmanaged(           std::string_view name);
+		void   Register         (TypeID ID, std::string_view name);
+		TypeID Register         (           std::string_view name);
 
 		// unmanaged
 		// non-const, non-volatile
@@ -323,6 +325,10 @@ namespace Ubpa::UDRefl {
 		TypeID RegisterAddConstLValueReference(TypeID ID);
 		TypeID RegisterAddRValueReference(TypeID ID);
 		TypeID RegisterAddConstRValueReference(TypeID ID);
+
+	private:
+		using IDRegistry<TypeID>::Register;
+		using IDRegistry<TypeID>::IsRegistered;
 	};
 }