+++ include/UDRefl/Object.h
@@ -17,6 +17,13 @@ bool operator op (Arg&& rhs) const {                                            
     return static_cast<bool>(ADMInvoke(StrIDRegistry::MetaID::operator_##name, std::forward<Arg>(rhs))); \
 }
 
+#define OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR(op, name)                             \
+template<typename Arg>                                                             \
+ObjectPtr operator op (Arg&& rhs) const {                                          \
+	AInvoke<void>(StrIDRegistry::MetaID::operator_##name, std::forward<Arg>(rhs)); \
+	return *this;                                                                  \
+}
+
 #define OBJECT_PTR_DECLARE_CONTAINER(name) \
 template<typename Arg>                     \
 SharedObject name (Arg&& rhs) const
@@ -29,10 +36,11 @@ SharedObject name (Args&&... args) const
 template<typename Arg>                     \
 SharedObject operator op (Arg&& rhs) const
 
-#define SHARED_OBJECT_DEFINE_OPERATOR(op)                       \
-template<typename Arg>                                          \
-SharedObject operator op (Arg&& rhs) const {                    \
-    return AsObjectPtr()->operator op (std::forward<Arg>(rhs)); \
+#define SHARED_OBJECT_DEFINE_OPERATOR(op)                \
+template<typename Arg>                                   \
+SharedObject operator op (Arg&& rhs) const {             \
+    AsObjectPtr()->operator op (std::forward<Arg>(rhs)); \
+    return *this;                                        \
 }
 
 #define SHARED_OBJECT_DEFINE_CMP_OPERATOR(op)                   \
@@ -405,6 +413,10 @@ namespace Ubpa::UDRefl {
 			Args&&... args) const;
 
 		// A means auto, ObjectPtr/SharedObject will be transform as ID + ptr
+		template<typename T, typename... Args>
+		T AInvoke(StrID methodID, Args&&... args) const;
+
+		// A means auto, ObjectPtr/SharedObject will be transform as ID + ptr
 		template<typename... Args>
 		SharedObject AMInvoke(
 			StrID methodID,
@@ -435,16 +447,16 @@ namespace Ubpa::UDRefl {
 			std::enable_if_t<!std::is_same_v<std::remove_cv_t<std::remove_reference_t<Arg>>, ObjectPtr>, int> = 0>
 		SharedObject operator=(Arg&& rhs) const;
 
-		OBJECT_PTR_DECLARE_OPERATOR(+=, assign_add);
-		OBJECT_PTR_DECLARE_OPERATOR(-=, assign_sub);
-		OBJECT_PTR_DECLARE_OPERATOR(*=, assign_mul);
-		OBJECT_PTR_DECLARE_OPERATOR(/=, assign_div);
-		OBJECT_PTR_DECLARE_OPERATOR(%=, assign_mod);
-		OBJECT_PTR_DECLARE_OPERATOR(&=, assign_band);
-		OBJECT_PTR_DECLARE_OPERATOR(|=, assign_bor);
-		OBJECT_PTR_DECLARE_OPERATOR(^=, assign_bxor);
-		OBJECT_PTR_DECLARE_OPERATOR(<<=, assign_lshift);
-		OBJECT_PTR_DECLARE_OPERATOR(>>=, assign_rshift);
+		OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR(+=, assign_add);
+		OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR(-=, assign_sub);
+		OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR(*=, assign_mul);
+		OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR(/=, assign_div);
+		OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR(%=, assign_mod);
+		OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR(&=, assign_band);
+		OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR(|=, assign_bor);
+		OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR(^=, assign_bxor);
+		OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR(<<=, assign_lshift);
+		OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR(>>=, assign_rshift);
 
 		OBJECT_PTR_DECLARE_OPERATOR([], subscript);
 		OBJECT_PTR_DECLARE_OPERATOR(->*, member_of_pointer);
@@ -820,6 +832,7 @@ namespace Ubpa::UDRefl {
 
 #undef OBJECT_PTR_DECLARE_OPERATOR
 #undef OBJECT_PTR_DEFINE_CMP_OPERATOR
+#undef OBJECT_PTR_DEFINE_ASSIGN_OP_OPERATOR
 #undef OBJECT_PTR_DECLARE_CONTAINER
 #undef OBJECT_PTR_DECLARE_CONTAINER_VARS
 #undef SHARED_OBJECT_DECLARE_OPERATOR